Implement hanging detection Make hanging detection also handle exception cases Remove tmp file "NUL" Fix so that rainbow don't skip outputting last character in all output

+++ extra_functions.py
@@ -27,7 +27,8 @@ class FileHelper:
 
     @staticmethod
     def remove_file(path_and_file):
-        os.remove(path_and_file)
+        if FileHelper.does_file_exist(path_and_file):
+            os.remove(path_and_file)
 
 
 class CommandHelper:
@@ -59,7 +60,7 @@ class CommandHelper:
 
     @staticmethod
     def strip_coloring(line):
-        matches = re.findall('\[[0-9]{1-2}m?', line)
+        matches = re.findall('\[[0-9]{1,2}m?', line)
         if matches is not None and isinstance(matches, basestring):
             matches = [matches]
         if matches is not None and len(matches) > 0:
@@ -111,10 +112,9 @@ class ColorOutput:
                                     ColorOutput.GRAY if color == 'gray' else ColorOutput.BLACK)))))))
                 return color_prefix + text + ColorOutput.END
             else:
-                text = CommandHelper.strip_coloring(text)
                 output = ''
                 offset = ColorOutput.rainbow_color_offset % len(ColorOutput.RAINBOW_LIST)
-                for i in range(0, len(text) - 1, 1):
+                for i in range(0, len(text), 1):
                     if text[i] == ' ':
                         offset += 1
                         output += ' '
